requirements.txt
Flask==3.0.0
pytesseract==0.3.10
Pillow
PyMuPDF==1.23.8
pypdf==4.0.1
docx2pdf==0.1.8
openpyxl==3.1.2
python-docx==1.1.0
Werkzeug==3.0.1
boto3
gunicorn
Babel==2.15.0
psycopg2-binary